Add tests and fix svg cases for non image magik deploys

+++ plugins/entando-plugin-jacms/src/main/java/com/agiletec/plugins/jacms/aps/system/services/resource/model/ImageResource.java
@@ -172,7 +172,7 @@ public class ImageResource extends AbstractMultiInstanceResource {
                 }
             }
         } catch (ImageProcessingException|IOException ex) {
-            logger.error("Error reading image metadata", ex);
+            logger.error("Error reading image metadata for file {}", file.getName(), ex);
         }
         return meta;
     }
@@ -249,13 +249,20 @@ public class ImageResource extends AbstractMultiInstanceResource {
                 //yes so configure the path where ImagicMagick is installed
                 convertCmd.setSearchPath(this.getImageMagickPath());
             }
-            // create the operation, add images and operators/options
-            IMOperation imOper = new IMOperation();
-            imOper.addImage();
-            imOper.resize(dimension.getDimx(), dimension.getDimy());
-            imOper.addImage();
-            convertCmd.run(imOper, bean.getFile().getAbsolutePath(), tempFile.getAbsolutePath());
-            this.getStorageManager().saveFile(subPath, this.isProtectedResource(), new FileInputStream(tempFile));
+
+            //svg files won't resize correctly via the image processor so save them directly
+            if(bean.getMimeType().contains("svg")) {
+                this.getStorageManager().saveFile(subPath, this.isProtectedResource(), new FileInputStream(bean.getFile()));
+            }else {
+                // create the operation, add images and operators/options
+                IMOperation imOper = new IMOperation();
+                imOper.addImage();
+                imOper.resize(dimension.getDimx(), dimension.getDimy());
+                imOper.addImage();
+                convertCmd.run(imOper, bean.getFile().getAbsolutePath(), tempFile.getAbsolutePath());
+                this.getStorageManager().saveFile(subPath, this.isProtectedResource(), new FileInputStream(tempFile));
+            }
+
             boolean deleted = tempFile.delete();
 
             if (!deleted) {
@@ -273,14 +280,21 @@ public class ImageResource extends AbstractMultiInstanceResource {
             // skips element with id zero that shouldn't be resized
             return;
         }
+
+
         String imageName = getNewInstanceFileName(bean.getFileName(), dimension.getIdDim(), null);
         String subPath = super.getDiskSubFolder() + imageName;
+
         try {
-            this.getStorageManager().deleteFile(subPath, this.isProtectedResource());
-            IImageResizer resizer = this.getImageResizer(subPath);
-            ResourceInstance resizedInstance
-                    = resizer.saveResizedImage(subPath, this.isProtectedResource(), imageIcon, dimension);
-            this.addInstance(resizedInstance);
+            if(bean.getMimeType().contains("svg")) {
+                this.getStorageManager().saveFile(subPath, this.isProtectedResource(), new FileInputStream(bean.getFile()));
+            }else {
+                this.getStorageManager().deleteFile(subPath, this.isProtectedResource());
+                IImageResizer resizer = this.getImageResizer(subPath);
+                ResourceInstance resizedInstance
+                        = resizer.saveResizedImage(subPath, this.isProtectedResource(), imageIcon, dimension);
+                this.addInstance(resizedInstance);
+            }
         } catch (Throwable t) {
             logger.error("Error creating resource file instance '{}'", subPath, t);
             throw new ApsSystemException("Error creating resource file instance '" + subPath + "'", t);
